chore: Rename runtime/pipeline.rs → tool_exec.rs; move command/pipeline… [ORB-10391]
Rename runtime/pipeline.rs → tool_exec.rs; move command/pipeline_run.rs → command/job/pipeline.rs

Planned-By: codex
Agent-Model: gpt-5.6-terra
