+++ crates/orbit-core/src/command/tests/mod.rs
@@ -1,3 +1,3 @@
 mod executor;
+mod job_pipeline;
 mod learning_authoring;
-mod pipeline_run;
