+++ crates/orbit-core/src/runtime/tests/mod.rs
@@ -1,8 +1,8 @@
 mod builder;
-mod pipeline;
 mod resolve;
 mod run_audit;
 mod run_input;
 mod runtime;
 mod task_block_on_run_failure;
 mod task_reservation_cleanup;
+mod tool_exec;
